Fixed a small bug

+++ ldpred/reporting.py
@@ -6,8 +6,8 @@ def print_summary(summary_dict, title, log_file=None):
     print('')
     title_len= len(title)
     num_dashes = 80-title_len-2
-    left_dashes = '-'*(num_dashes/2)
-    right_dashes = '-'*(num_dashes/2+num_dashes%2)
+    left_dashes = '-'*int(num_dashes/2)
+    right_dashes = '-'*int(num_dashes/2+num_dashes%2)
     
     print( left_dashes  + ' '+title+' '+right_dashes)
     for k in sorted(summary_dict.keys()):
